feat: add certified classics
